save-to-workbench/requirements.txt
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==2.18.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.0.0
requests==2.32.5
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
